Paper pages: New mode selector, similar to the older mode selector.
Show modes as images with textual captions, above the left-hand overhang
(so somewhat like old-style tabs).  To be reconsidered; the other style,
with images and text towards the right side of the title area, is still in
the code.

Also update the CSS to avoid overhang problems when the paper table is too
wide.

Also add a review image different from the edit image.