Tag setting: Correct error that could corrupt tags.
The PHP "foreach (... as &$var)" is very dangerous, because after the loop
is over, $var is still bound to a reference, and setting that variable
again -- even in another "foreach (... as &$var)" loop -- will affect the
referred object.

The visible effect of this problem was that an added tag was ignored,
because the added tag was overwritten with the first tag in the list.