Add chair-only tags, such as "~~tag".
A tag with two twiddles, such as "~~tag", is only visible, to and
changeable by, chairs and system administrators only.

Andrew Myers requested this feature.  He implemented it with an initial
underscore: "_tag" is a chair-only tag.  Haven't decided which is better,
but the "~~tag" implementation should be in the history.