Detect and report max_allowed_packet problems.
It turns out that most (all?) MySQL versions either ignore or reject
per-session settings of max_allowed_packet.  So restore the README text
that tells the user to set max_allowed_packet, and add an admin-only test
that max_allowed_packet is big enough to accept a paper.

John Heidemann reported a bug that might be explained by this problem.