Account creation improvements: Support creating accounts on LDAP settings.
While the better thing would probably be to extract information from LDAP
when the account is created, to for example ensure the account exists, we
don't do that yet.

Also simplify HTML coding, verify that the preferred email is valid, and
keep track of list position even after a form submission.

Problem reported by Gareth Gale.