Paper edit screen: Update messaging.
Hopefully this will pacify Jeff Mogul's anxious users, who were asking if
they could update after upload.  The answer was yes, and if they had
uploaded, a message would have clearly told them so; but the message on the
"New paper" screen was a bit ambiguous.  It now should read:

  Enter information about your paper. You can make changes until the
  deadline, but thereafter incomplete submissions will not be
  considered. The deadline is ....

which should be good enough.