Search internals: Refactor to a more convention expression style.
In other words, rather than creating an "array of clauses", create a single
expression, which might be an "and" or "or" expression with
explicitly-linked subexpressions.  Introduce support for "or" clauses.

This will help us much more naturally support parenthesized expressions and
such, and is simpler to think about.  Not that the code as is is simple
even now.

Also fix some bugs and change some behavior.
- "round:R2 cre:1" didn't work -- it found ANY review, not just completed
  reviews.  Now it works.
- There were some others.