Add "THEN" searches and multiple order: searches.
"THEN" is the lowest precedence search.  It is like "OR", but can only
appear at the top level, and also affects the sort order -- in a search
like "a THEN b", the papers matching "a" will appear in the list before the
papers matching "b".

"order:a THEN order:b" will also work; it searches for tag "a", and outputs
results as by "order:a", then searches for tag "b" sorted as by "order:b".
Papers that match both tag:a AND tag:b will be sorted first by order:a,
then by order:b.

This is intended to produce dynamic discussion orders, as in "order:discuss
THEN order:accept THEN order:reject".  Geoff Voelker requested dynamic
discussion orders.