Paper search: Move to less error-prone API for specifying a contact ID.
Sometimes we pass the whole request as the "options" field, so if there was
a "c" parameter, we'd treat that as a contact ID (on chair requests; useful
for review preference & manual assignment).  This seems risky and
unnecessary, so make the contact ID a separate argument.  There is a small
chance that someone passes "c" parameters intentionally; if so oh well,
I'll get a bug report eventually.