Replace HotCRP display of current time with live deadline display.
The upper right corner reports something like "Paper registration deadline
in 8 days", "Paper submission deadline in 30 seconds", "Paper submission
deadline is NOW" in advance of the deadline.  Javascript keeps this display
up to date.  Ajax loads are used to update deadlines and keep the browser's
clock in sync with the server's.  Grace periods are handled correctly;
"Paper submission deadline is NOW" is reported until the grace period is
over.

This seems better, cleaner, and more active than displaying the page load
date and time in the server's time zone.