Fix negated search terms again.
I remember now the purpose of this code (and have documented it).

A search like "abstract:bar" searches for THE WORD bar, not the substring
bar.  We do not, however, implement word boundaries in MySQL.  Instead we
load the abstract into PHP and use PCRE there to check for word boundaries.
This is more flexible.  The MySQL portion of the query checks for THE
SUBSTRING bar, and thus returns more papers than the precise answer.  This
is fine EXCEPT underneath a negation.  Consider the search "NOT
abstract:bar".  This matches any paper without THE WORD bar in the
abstract.  This INCLUDES some papers with bar as a SUBSTRING in the
abstract.  Therefore, including in SQL some term like ".. abstract not like
'%bar%' .." would be incorrect.

Negated search terms weren't quite handled correctly before in complex
cases, though.  They should be now.