Correct inappropriate exposure of review scores via search.
As a convenience feature, simple searches like "re:deborah" display review
information about Deborah's review -- the "reviewer type" icon describes,
and links to, Deborah's review, and the score graph highlight refers to
Deborah's score.

Unfortunately, this convenience feature was implemented internally by
changing the "identity" of the user doing the search.  That is, the search
was performed *as if by Deborah*.  As a result, a PC member searching for
"re:deborah" was shown the results *that Deborah would see*, rather than
the results that the PC member should see.

Note, however, that information about conflicted papers was NOT exposed
(because conflict information was correctly tied to the PC member).  The
exposure bug only affected non-conflicted papers.

The correction is to make sure that fields like "myReviewType", etc. always
refer to the requesting contact.

I doubt this bug has been exposed for too long -- some recent fixes to
search may have made it more likely (for instance
9a8d9d1a03765b4fe0394f8fbf530c209687b1eb) -- but am not sure.

Aditya Akella/Jeff Mogul report.