Paper capabilities: Initial implementation.
A paper capability is a generalization of a review token, and gives the
right to do something to a paper, independent of the current contact's
explicit rights as encoded in the database (or whether, in fact, the
current contact is validated).  The initial implementation gives users the
right to view a specific paper as an author of that paper.  I'm only
testing the feature now; you set $Opt["testCapabilities"] to true, and then
supply URL parameters like "testcap=4a,5a" to get _A_uthor view
capabilities on papers 4 and 5.  ("testcap=-4a" drops capability 4a.)  This
works even if the current user is not valid.

Some work remains.  Currently, an author-view capability makes the user
into a sort of semi-author -- a PC member, for example, can see a mixture
of author-visible stuff and PC-visible stuff.  Really, what's desired is
prioritization -- for instance, the author-view capability matters only if
the currently signed in user couldn't see the paper otherwise.  That
remains to be implemented.