Correct problem with review versions in multi-form downloads.
John Wilkes reported this problem -- a sort of off-by-one error, due to the
fact that "==+== Version" lines precede the "==+== Paper" lines that
previously caused HotCRP to stop parsing one form and begin parsing the
next.  This meant that the "==+== Version" for review X+1 might actually be
applied to review X.

The solution is to stop parsing one form when encountering "==+== Begin
Review".