Add filestore support.
The filestore is a cache, on the local filesystem, of the paper contents in
the database. The idea being that it is faster and easier on the server to
send data from the local filesystem than to read data from the database,
through PHP.

The primary location for paper data is still the database; blowing away the
filestore should be safe. This seems cleaner; for example, the SQL dump
created by Code/backupdb.sh is complete.

*Modifying* filestore contents is not safe.

If $Opt["filestore"] is true, then HotCRP will attempt to create a
filestore in INSTALLATIONDIR/filestore. If $Opt["filestore"] is a string,
then HotCRP will use that directory.

The filestore layout is FILESTORE/SHA1PREFIX/SHA1.EXT, where SHA1 is the
file's SHA-1 sum in hex, EXT is the relevant extension (e.g. "pdf"), and
SHA1PREFIX is the first three characters of SHA1.