Correct unfortunate review token bug reported by Kevin Fu.
Imagine a conference where review tokens are used for anonymity for chair
reviews. The "offline review" page was dangerous. Anyone COULD upload a
review for the paper; the review would be given their name. Too bad, this
happened in exactly this way *even for users who had the review token*. Now
we check for a relevant review token when deciding which review to replace.
Blech.