Contact an updates server and report necessary updates.
I don't want the next vulnerability to catch people by surprise. So on the
server side, detect the current git version numbers and pass them to the
user's browser. On the browser side, pass these numbers to an updates
server and examine the response. The response can contain messages ("your
version has a problem") and conditions ("except if $Opt[x] == y"). If the
response is nonempty, pass it back to the HotCRP server for checking. Print
a message if the response says the problem is real. Only check once every
20 seconds until something is wrong. The default updates server is
hotcrp.lcdf.org/updates, but you can set it to something different (or turn
it off) with $Opt["updatesSite"].