Filestore security: Always check for .htaccess.
Previously, we added a .htaccess file to the filestore when CREATING the
filestore. But it seems likely that some users will do a "mkdir
filestore" and then run HotCRP without creating .htaccess. It would suck
to leave the resulting filestore directories unprotected, so check for
.htaccess every time.

This commit is actually redundant in most HotCRP installations, because
the HotCRP parent directory's .htaccess file disallows filestore access
using mod_rewrite. But better safe than sorry.