Progress on per-paper "managers."
A manager is someone who can act like the chair for a paper.

This commit:
* Adds a SQL field, Paper.managerContactId, for the manager.
* The user with ID P.managerContactId can administer the paper.
* If P.managerContactId!=0 and the chair has a conflict with P,
  then the chair CANNOT administer P (and can't directly override
  their conflict).

It will be extremely painful to verify that a conflicted chair can't
discover any info about a managed paper (e.g. through the assignment
process, through search). I doubt that verification process will ever
complete. This is a start.