Internal: Assigner update.
An AssignmentState object encapsulates the current set of
assignments for all papers. We prepare an assignment by making
changes to the proposed state, which starts out as a copy of the
current state. We then prepare a diff between the current state
and the proposed state, and show that. This has a couple
advantages. For instance, we could allow users to upload a bulk
assignment that *replaces* the current assignment.

Also fix some bugs. Create new users outside LOCK TABLES.
Clean up some printouts. Etc.