Initial support for meeting navigation.
Meeting navigation is a follow-the-chair mode. The chair's
browser's current paper is stored on the server, and transmitted
to everyone else.

When meeting navigation is enabled, a bar appears at the top
of each screen with (1) the current paper, (2) the next paper,
(3) the paper after that. Links are provided. If you're not
looking at what the chair's looking at, the bar is red. If you
are synched with the chair, the bar is yellow.

There's no way to turn on navigation in the UI. A chair can
turn it on with URL editing: append "?nav=1" to the end of a
URL.

Required a bunch of changes and inspired some others.
- "list=" parameters become "ls=" parameters; e.g.,
  "ls=1+2+3+4".
- "ls=" parameters can take a list ID, such as "ls=p/s/foo/X"
  where X is a sort definition.
- PaperList class can "render" a sorted list of paper IDs
  when given the "idarray" => true option.
- Correct a bug in bd1e1415: we weren't ever saving a newly
  created session list.

Signed-off-by: Eddie Kohler <ekohler@gmail.com>