Be more careful when parsing times with timezones.
Consider the timezone abbreviation "IST" for Asia/Kolkata.
Unfortuately, the PHP strtotime() function, when it encounters
a timezone abbreviation, uses the first matching timezone
definition in the timezone database, which for IST is
Asia/Jerusalem!, rather than preferring the timezone that equals
the current timezone, if there is one.

And abbreviations like GMT-5 (for Etc/GMT-5) aren't even parsed.

Work around this with regular expressions.

Reported by Aakash Sharma.