Support long polling in the meeting tracker.
Run `node src/trackercomet.js`, and configure HotCRP with
$Opt["trackerCometSite"] = the site for src/trackercomet.js.
Then the browser does a long poll to the trackercomet server,
passing the current "tracker status". The current tracker status
is just a "tracker ID" plus the last time the tracker changed
status, so it's safe to expose publicly. When the long poll
expires, the browser goes back to the normal deadlines URL.