Reduce surprise in quicksearch.
When searching for a term that returns a unique result, where that result
is in the current list, preserve the current list *only* if the current
list is interesting.

Probably an even better answer is to always make a new list but I do
remember why the current behavior is useful.