Support "combine" graphs, which aggregate over multiple papers.
Combine formulas have a track phase and a combine phase. In the
track phase, you compute the components of each aggregate and
produce an array of those components. In the combine phase,
you compute the aggregate from all those components.

Create a combine graph with "bar FORMULA". "count" and "fraction"
graphs are retconned to combines.

The mental model here needs work before it can be explained to anyone.