Password improvements.
The `passwordIsCdb` field is a bad idea. The problem is that now we
can have a valid-looking password in the DB that is not in fact
valid. The main change here is to improve plaintext_password() to
check passwordIsCdb and delegate to the contactdb password. Really
we should get rid of passwordIsCdb.