Fix an information exposure bug in tag setting.
The assignment framework is used for tag setting now. This is
generally good, but it meant a clever PC member could find out
whether or not a tag existed for a conflicted paper by trying
to remove it! Specifically, say the PC user wanted to see if
conflicted paper #5 had tag #foo. They could (using search actions)
"remove #foo" from #5. In the assignment framework, this works by
(1) loading the current set of tags, (2) recording all differences
between the current set of tags and the desired set of tags, (3)
checking permissions on any differences. In particular, if there
are no differences, there's no permissions to check. So if #foo
was on 5, they'd get an error message. If #foo wasn't on 5, they'd
get no error message!

Try to fix this while preserving useful properties of the assignment
framework. Specifically if you try to remove "#foo*" you get no error
message either way.