Support session list digests.
In a session list digest, the list is stored client side, and
referenced by a digest name. This could be crypto but because the
browser crypto API is so bleh it is not, it's just a timestamp.

If the list appears so long as to overflow a cookie, don't store
the IDs in the cookie, instead send a digest and parse the response.

Fucking fuck fuck shit fuck fuck fuck fucking