a backup script to send backups over email when changed
This is a home-grown backup script that I built for a local HotCRP
installation, but I thought that it may be useful to others as
well. The backup strategy is as follows:

- a backups/ directory at the project root with at most two backups saved,
  the last one (`new`) and the one before that (`old`)

- whenever the script runs, it moves `new` to `old` and creates a new
  `new` file using the `backupdb.sh` script; then it compares `new`
  and `old`, and sends the new backup by email if it differs from the
  old one

- if a `passphrase` file exists in the `backups` directory, then
  it is used as a passphrase for `gpg` to encrypt the backup file

Finally, note that we are careful to pass `--skip-dump-date` to
`mysqldump`, otherwise the backups would always differ.

The script is careful to `cd` to its base directory, so that it can be
invoked directly from a crontab.