reviewStatus: Delete status type RS_APPROVED.
It is simpler and better to say that all submitted reviews,
including external reviews, become "completed", and do not disappear
from author view if an administrator changes the pc_editdelegate
setting to "Yes, and external reviews are visible only to their
requesters".

But update the pc_editdelegate setting UI to report if there are
submitted external reviews, and to link to a list of them.