Xor_SearchTerm: Check trivial_rights.
`x XOR y` means `(x AND NOT y) OR (y AND NOT x)`. So negation is
an inherent part of the XOR connector. That means it's very
important to check trivial_rights, because trivial_rights says
whether the SQL generated by `sqlexpr` is safe to negate.

This issue may explain why SODA 21 vote counts went off.