Deprecate Qrequest::post_ok() in favor of valid_post().
valid_post() checks for the POST method as well as the CSRF token.

And add a check for the HEAD method to uses of URL capabilities for
review accept and decline. It looks like some of the weird behavior
I'm seeing is due to MS Azure calling HEAD methods on those URLs.
GET methods are "not supposed to change state", but HEAD methods
definitely should not.