Fix handling of parenthesized text in AbbreviationMatcher.
The xtester was removing parentheses altogether. This prevented
it from matching strings created in _xfind_all, which replaces
parentheses with spaces (effectively).

(It is an invariant (requirement) that the matcher expression
created for string S in AM::_xfind_all MUST match the xtester
created for string S by AM::make_xtester(S).)