Fix submit-on-first-save.
I noticed that if you fill out a complete paper and submit it,
with the "Ready for review" checkbox checked, the paper still
ends up a draft. This was because paper.php looked for
`has_file("opt0")` rather than `has_file("submission")`, and
in 183a440cf467b73ad77871f9b26eedf10f3f4f41, we changed the
file upload name to "submission". It's safer to do things
more this way; the `submitpaper` name always means "submit this."