Meeting tracker: Per-tracker setting of "Hide conflicted papers".
And internal rewrite of MeetingTracker to rely on true objects,
rather than arbitrary JSONs. As with many rewrites for clarity this is
unwise but I couldn't take it any more.

And turn on hide-conflicted-paper-IDs by default.

And if hide-conflicts is on for a tracker, remove conflicted paper IDs from
the relevant list.