Introduce placeholder users.
A placeholder user is effectively not there; the record is a
placeholder for a user who has never logged in or been explicitly
created. A placeholder user is "dormant" (should not show up in
lists or be sent email) but not "disabled".

The hope is that placeholder users will reduce the need for funky
email analysis (e.g., "email_authored_papers"); keep the contactdb
roles more actively up to date; and simplify multi-account
handling.

Tests are updated to account for new placeholder user behavior.