Accept/decline workflow does not use magic GET links.
The link previews implemented by Outlook do interpret Javascript,
thereby clicking the POST redirect form that we relied upon.
Perhaps if we called `click()` on the submit button rather than
`submit()` it would work OK? But this is tiresome.

So instead of accept/decline links, add a big accept/decline banner
to the top of the relevant pages until the review is begun.

* Update look of “special link” messages.
* Update mail.