Add messages saved by token; use to avoid API redirection.
Previously, some APIs, such as `acceptreview`, could take a `redirect`
parameter. These APIs would respond to a POST not by returning JSON,
but by redirecting the user's browser to a location defined by the
user.

Now, these APIs instead store messages under a named token in the user's
session, and Javascript is used to reload the page with that token,
thus showing the messages.