Great extension parameter internals rewrite.
When I am going to get anything right?
The old method for handling extension checkers, which involved
global variables like $conf->xt_checker and $conf->xt_allow_function,
caused a difficult failure in conferences with complex requirements.
Specifically, a user tried to download reviews (via a ListAction)
from a conference with a complex review-visibility requirement
(namely, a response round condition).
ListActions have constrained availability based on the contents of
a list. So ListAction sets a relevant checker (to support things
like `"display_if": "listreport:pf"`), and an allow function (to
ensure that XSS checking occurs). But one ListAction checks whether
the viewing user can view any review. Which instantiates the
current response rounds' search conditions. If those were already
created for some reason, great, no problem. If not, then those
search conditions were parsed. But that's a problem because the
parsing of search conditions inappropriately used the allow function
intended for ListActions!!!!!

So introduce an XtParams structure and get rid of the globals.

This is an incompatible change.