Get rid of "general" OVERRIDE_TIME functionality.
It was odd for administrators to be subject to deadlines only when
a specific flag (OVERRIDE_CHECK_TIME) was set. This caused a
problem with withdrawing a paper at ASPLOS under an uncommon
setting regime. So remove that functionality. In display mode, use
author_user() to check whether withdrawing, editing, etc. is
actually allowed under the deadlines; this is more robust.