Chair reviewers are not sent notifications on incomplete reviews.
An administrator can always see reviews for a paper, even if their
own assigned review for that paper is incomplete. But it is a
little too aggressive to send the administrator NOTIFICATIONS for
submitted reviews, etc. So don't do that.

Requires differentiating followers for the paper's *reviews*
from followers for the *paper* (e.g. comments with topic paper).