Add sort:color.
Interpolate the colors corresponding to a paper in the CIEHcl color
space, then compare hue, lightness, and chroma.

"Good" behavior is not well-defined. It's not smart to compare
hue, lightness, and chroma only, because colors in a ramp often
differ by small amounts in hue even if they "should" have the
same hue. So compare first by quantized hue, lightness, and
chroma, and only then by fine-grained hue, lightness, and chroma.