Fix creating docstore
If `docstore` is set to true in the options.php, hotcrp fails to create the `docs` directory because in `_make_fpath_parents` `$container` will be `<root>/docs` and `$fdir` is `<root>/docs/`. Therefore, the check in line 470 `strlen($container) < strlen($fdir)` will be true and the function returns without creating the directory.

This fix removes trailing `/` from `$fdir` as it is done for `$container` and solves this problem.