Autogrow nit.
On initial load, the trick where we change an element's width
to 16px and then examine its scrollWidth does not work. The
consequence of this is that we add in extra padding, but not
for empty elements (which are filtered out).

So do not change width if the scrollWidth fits in the initial
element width.