Avoid self-DOS when there's a tracker open for an `actas` user.
Since the `actas` user can't see the tracker, it gets eventid: 0.
If the other user as eventid: NONZERO, then the tab with `actas`
wars with the other tabs over the correct value for trevent
storage, resulting in a reload storm.

Solution: different storage locations for different users.
AGH

Also standardize on `hotcrp.status`.