Remove the sf/*/presence setting; use only .../condition.
Also compute `PaperOption->phase` (which replaces `->final`)
from the exists_condition. This avoids situations where the
exists_condition and the phase are out of sync -- e.g., if the
user specifies a condition like `phase:final #hello`. That can
be problematic because the exists_condition() is computed for
the conference root user, who can always see all decisions;
but final versions should only exist when the VIEWING user
can see the decision.