`any`, `all`, `none` searches are always special regardless of case.
Before there was some "logic" so that `ALL` was special, but `all`
was only special near a `THEN`. This was stupid and error prone and
caused some confusion with automatic tags, which add an implicit
`searchoption:expand_automatic` operator to the search (therefore
making `all` not alone or near a `THEN`).

It might be nice for `any`/`all`/`none` to LOSE their specialness
in a context like `any of the above` that's near other barewords,
but keeping them caselessly special is much better default.