Add a random `name` attribute to the paper form.
On reloading a page, Firefox fills in each control with the value
of "that control" from before the reload. Unfortunately, "that
control" is defined according to *order* in the form, not, e.g.,
`name` of control. That means that if, for example, the PC has
changed since the initial load, PC conflicts may be displaced!

Example: Initial PC conflicts:
   pcconf:1 [X]     pcconf:2 [ ]     pcconf:3 [X]     pcconf:4 [ ]
User 2 is removed from the PC; author reloads. Expected state:
   pcconf:1 [X]     pcconf:3 [X]     pcconf:4 [ ]
Actual state on FF (which fills in controls based on form order):
   pcconf:1 [X]     pcconf:3 [ ]     pcconf:4 [X]

Two PLDI2024 papers experienced this problem. A solution from
Stack Overflow is to set `autocomplete="off"` on the form. That
seems heavyweight and has additional consequences, e.g., preventing
browsers from filling in form entries on *history navigation*
(which is unproblematic). Another SO solution is to give the form
a random `name`, which appears to work.