banal: Address #329 by keeping measurements with more precision.
Instead of rounding measurements to the closest integral number of
points (1/72" points), round to the closest EIGHTH of a point. The
resulting gain in precision lets us, as expected, round 11.955pt
up to 12pt.

Rounding and binning always has anomalies, though; if a more-precise
leading is 1.75x less likely than an integral leading, use the
integral leading.

Also:
* Remove unused banal_leading_policy.
* Coerce more values to numerics.
* Close gaps in columns that are due to really narrow columns.