Contact::session_users() can contain empty strings.
Previously, if a user was signed out of a single account (via
LoginHelper::change_session_users), that would shift all higher
account indexes down. This seems wrong since, in the context of
that session, other tabs rely on those indexes implying specific
users.

The commit causes a user signing out of a single account to replace
that index in `us` with an empty string. The `initialize_user` logic
handles those kinds of users.

This is all a huge pain in the ass, and likely we should not allow
single accounts to sign out of a session, or should change session
cookies if and when that occurs. :(