Tag <a> elements are auto-populated on use.
Maybe someone is relying on the `href` attributes being real, we'll
see. This is nicer because (1) the HTML is smaller, (2) we can
decide from context what to link to -- for example, a tag on the
review preferences page can link back to the review preferences
page (#332), and a tag on a search page will incorporate any view
annotations on the search.