Introduce new mode for review search.
A search like `Field:none` should mean the same as `NOT Field:any`.
It did not mean this, because it meant instead something like
“has a submitted review that has no Field”.

If ReviewSearchMatcher::$mode === 1, implement better behavior,
where `Field:none` means `NOT Field:any`. Also implement `Field:blank`,
which means “the Field exists, but its value is blank.”

This mode will be the default but first want to know what changes.