Internal: PaperContactInfo readability.
- Add is_author, is_reviewer, conflicted, unconflicted.
- Rename allow_review to allow_reviewer.
- Remove potential_reviewer (1).
- Replace allow_reviewer with is_reviewer (2).

(1) The potential_reviewer flag was used in only two places that
determined permissions, and it oddly checked the self-assignment
track but not the ability to self-assign. In those places
(can_view_manager + can_view_conflicts) it's better to check
is_reviewer + allow_pc[_broad].

(2) allow_reviewer in combination with allow_pc is redundant,
because allow_pc is a superset of allow_reviewer with the exception
of external reviewers.

Also remove some obsolete checks: view_score_bound $rrow argument
is required.