DocumentInfoSet: README-warnings.txt has stable timestamp
Giving README-warnings.txt the current time as a timestamp is a problem,
since the content-signature (which didn't include the current time)
would then not reflect the actual content of the ZIP file.

So give README-warnings.txt the time of Anne Dudfield's birthday by
default. And include timestamps in the content-signature.