Depressingly large graph changes
It was not intuitive that a graph of `reviewer` vs. `avg(RevExp)`
produced a huge scatterplot, where the dots over reviewer R
indicated the average reviewer expertise *over all reviews* on
the papers R reviewed.

More intuitive is that the `avg` function refers to reviewer R's
reviews. So that is what this does.

If you want the former behavior, specify the data type with
`reviewer` vs. `paper avg(RevExp)`.

Also redo the graph JS to rely less often on unreadable integer
indexes.

Though this does make progress it feels bad.