Primary contacts don't get contact authorship
The previous version of primary contact IDs would grant the primary
linked account the same kind of authorship that a secondary linked
account had. For instance, if X became primary for Y, and Y was
contact author for #10, then X became contact for #10.

This was weird because if the link was broken (Y unlinked themselves
from X), then X *remained contact for #10*.

In this version, X becomes a *resettable* author for #10, and if Y
unlinks from X, then X's authorship is revoked.

Required a chunk of surgery and changes to some tests.