Change implementation of submission field conversions
Need support for optional conversions. The "conversions" member
of optiontypes.json is a list of objects, such as `{"from":"type"}`
or `{"to":"type"}`. The "allow_if" member, if set, is a callback
that returns truthy if the conversion is OK. The "setting_function"
member, if set, is a callback called if the conversion should
succeed.

And add test for real-number conversion (from numeric).

And fix value renumbering for checkboxes fields; we didn't handle
deleting values correctly.