Improve home-page display of "Review proposals" link
Contact::is_primary_administrator was broken because it did not
consider the possibility that a *conference chair* could also be
a track manager.

This should address an RTSS 2025 issue.

Also fix an inappropriate SQL expression for non-chairs, because
non-chairs may not be able to view the papaer manager (but they
can if it's them).