Introduce PaperOption::value_reconcile
This function is called to reconcile fields after modifications are
parsed. Its purpose is to ensure invariants across fields.
The return value of value_reconcile is a new PaperValue or null if
the old PaperValue is appropriate. value_reconcile is called on all
fields until fixed point, up to 10 times.